+++ crates/red_knot_python_semantic/src/lib.rs
@@ -29,6 +29,7 @@ pub(crate) mod symbol;
 pub mod types;
 mod unpack;
 mod util;
+mod visibility_constraints;
 
 type FxOrderSet<V> = ordermap::set::OrderSet<V, BuildHasherDefault<FxHasher>>;
