+++ crates/red_knot_python_semantic/src/semantic_index.rs
@@ -27,7 +27,6 @@ pub mod definition;
 pub mod expression;
 pub mod symbol;
 mod use_def;
-pub(crate) mod visibility_constraints;
 
 pub(crate) use self::use_def::{
     BindingWithConstraints, BindingWithConstraintsIterator, DeclarationsIterator,
