+++ crates/red_knot_python_semantic/src/types.rs
@@ -33,7 +33,6 @@ use crate::types::diagnostic::INVALID_TYPE_FORM;
 use crate::types::mro::{Mro, MroError, MroIterator};
 use crate::types::narrow::narrowing_constraint;
 use crate::{Db, FxOrderSet, Module, Program, PythonVersion};
-pub(crate) use static_visibility::analyze;
 
 mod builder;
 mod call;
@@ -45,7 +44,6 @@ mod infer;
 mod mro;
 mod narrow;
 mod signatures;
-mod static_visibility;
 mod string_annotation;
 mod unpacker;
 
@@ -275,16 +273,12 @@ fn bindings_ty<'db>(
         binding,
         constraints,
         all_constraints,
-        all_visibility_constraints,
+        visibility_constraints,
         visibility_constraint,
     } in bindings_with_constraints
     {
-        let static_visibility = analyze(
-            db,
-            all_constraints,
-            all_visibility_constraints,
-            visibility_constraint,
-        );
+        let static_visibility =
+            visibility_constraints.analyze(db, all_constraints, visibility_constraint);
 
         let Some(binding) = binding else {
             if !static_visibility.is_always_false() {
@@ -363,15 +357,11 @@ fn declarations_ty<'db>(
     let mut is_unbound_visible = false;
     let mut types = vec![];
 
-    for (declaration, all_constraints, all_visibility_constraints, visibility_constraint) in
+    for (declaration, all_constraints, visibility_constraints, visibility_constraint) in
         declarations
     {
-        let static_visibility = analyze(
-            db,
-            all_constraints,
-            all_visibility_constraints,
-            visibility_constraint,
-        );
+        let static_visibility =
+            visibility_constraints.analyze(db, all_constraints, visibility_constraint);
 
         let Some(declaration) = declaration else {
             if !static_visibility.is_always_false() {
